backend/requirements.txt
asgiref==3.2.10
Django==3.1.1
django-cors-headers==3.10.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.3.0
gunicorn==20.1.0
PyJWT==1.7.1
pytz==2021.3
sqlparse==0.4.2
psycopg2-binary==2.9.1
